+++ tests/src/testMaterial.cpp
@@ -103,7 +103,7 @@ auto testMaterialWithStrain(const MaterialImpl& mat, const double tol = 1e-14) {
 
   using ReturnTypeSecondDerivative = Ikarus::Std::ReturnType<decltype(ddf), decltype(ev)&>;
 
-  auto nonLinOp    = Ikarus::NonLinearOperator(linearAlgebraFunctions(f, df, ddf), parameter(ev));
+  auto nonLinOp    = Ikarus::NonLinearOperator(functions(f, df, ddf), parameter(ev));
   auto subNonLinOp = nonLinOp.template subOperator<1, 2>();
 
   t.check(checkGradient(nonLinOp, {.draw = false, .writeSlopeStatementIfFailed = true})) << "checkGradient Failed";